requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
idna==3.2
numpy==1.21.2
pandas==1.3.2
pandasql==0.7.3
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
